Move all 'global' CSS to frontend/src/css, move self contained CSS to .css file corresponding to React component
