+++ frontend/src/components/ColumnarSelector.css
