+++ frontend/src/components/DatabaseDetailsForm.jsx
