--- frontend/src/components/badge/badge.css
@@ -1,23 +1,0 @@
-:root {
-    --badge-border-radius: 4px;
-
-    --permissions-badge-color: #A989C5;
-    --headsup-badge-color: #F5A623;
-}
-.Badge {
-    border-radius: var(--badge-border-radius);
-    padding: 0.5em;
-    text-transform: uppercase;
-    font-weight: bold;
-    font-size: 0.6rem;
-    color: #fff;
-    background-color: #444;
-}
-
-.Badge.Badge--permissions {
-    background-color: var(--permissions-badge-color);
-}
-
-.Badge.Badge--headsUp {
-    background-color: var(--headsup-badge-color);
-}
