--- frontend/src/components/entity/entity.css
@@ -1,63 +1,0 @@
-.EntityField {
-  position: relative;
-  padding: 12px 0 24px;
-  color: #777;
-  border-bottom: 1px solid #f0f0f0;
-}
-
-.EntityField:hover {
-  background-color: #fafafa;
-}
-
-.EntityInfo {
-  margin-left: 64px;
-}
-
-.EntityFieldName {
-  font-size: 18px;
-  font-weight: 700;
-  color: #777;
-}
-
-.EntityDescription {
-  display: block;
-  color: #999;
-}
-
-.EntityVisibility {
-  margin: 17px 0 0 24px;
-}
-
-.EntityOriginalType {
-  margin-top: 11px;
-  margin-left: 11px;
-}
-
-.EntityCustomType {
-  margin: 14px 24px 0 0;
-}
-
-.EntityFieldHidden {
-  opacity: 0.4;
-}
-
-.CustomTypeApplied {
-  color: #4A90E2;
-  border-color: #4A90E2;
-}
-
-.EditedEntity {
-  position: relative;
-  color: #4A90E2;
-}
-
-.EntityField:hover .Drag-handle {
-  display: block;
-}
-
-.Drag-handle {
-  position: absolute;
-  top: 50%;
-  left: 5px;
-  margin-top: -21px; /* half the height of the item, which is 42px */
-}
