--- frontend/src/components/table/table.css
@@ -1,126 +1,0 @@
-:root {
-  --table-border-color: rgba(213, 213, 213, .3);
-  --table-alt-bg-color: rgba(0, 0, 0, .02);
-
-  --entity-image-small-size: 24px;
-  --entity-image-large-size: 64px;
-}
-
-/*
-  explicitly set the th text alignment to left. this is required for IE
-  which follows the suggested rendering and defaults to center, whereas
-  chrome and others do not
-*/
-th { text-align: left; }
-
-.Table {
-  /* standard table reset */
-  border-collapse: collapse;
-  border-spacing: 0;
-
-  width: 100%;
-
-  font-family: "Helvetica Neue", "Helvetica", sans-serif;
-  font-size: 0.76rem;
-  line-height: 0.76rem;
-  text-align: left;
-}
-
-.Table--bordered {
-  border: 1px solid var(--table-border-color);
-}
-
-.Table tr {
-  border-bottom: 1px solid var(--table-border-color);
-}
-
-.Table tr:nth-child(even) {
-  background-color: var(--table-alt-bg-color);
-}
-
-.Table th,
-.Table td {
-  padding: 1em;
-  border: 1px solid var(--table-border-color);
-}
-
-.EntityImage {
-  border-radius: 99px;
-  box-shadow: 0 0 4px rgba(0, 0, 0, .12);
-  background-color: #fafafa;
-}
-
-.EntityImage--small {
-  width: var(--entity-image-small-size);
-  height: var(--entity-image-small-size);
-}
-
-.EntityImage--large {
-  width: var(--entity-image-large-size);
-  height: var(--entity-image-large-size);
-}
-
-.EntityTableWrapper {
-  margin: 1em;
-}
-
-.EntityName {
-  margin-top: .25em;
-}
-
-.EntityTable tbody {
-  color: #555F6B;
-}
-
-.EntityTable td {
-  padding: 5px;
-  border: 1px solid #E7E7E7;
-}
-
-@media screen and (--breakpoint-min-lg) {
-  .EntityTable td {
-    padding: 10px;
-  }
-}
-
-.EntityTable .EntityTable--columnSelected {
-  color: #4A90E2;
-}
-
-.EntityTable tbody tr:hover {
-  background-color: #E3F0FF;
-}
-
-.EntityTable th {
-  color: #555F6B;
-  padding: 1em;
-}
-
-.EntityTable thead th:hover {
-  cursor: pointer;
-}
-
-.Message {
-  padding-left: 1.5em;
-  margin-top: 1em;
-}
-
-.Message-sender {
-  font-size: 0.8em;
-}
-
-.Message-text {
-  line-height: 1.5em;
-  max-width: 65%;
-  color: #666;
-  margin-top: .25em;
-}
-
-.Message--alt .Message-text {
-  color: #4A90E2;
-}
-
-.Timestamp {
-  color: #999;
-  font-size: 0.8em;
-}
