--- frontend/src/components/table_wrapper/table_wrapper.css
@@ -1,8 +1,0 @@
-:root {
-  --table-wrapper-border-color: #d9d9d9;
-}
-.TableWrapper {
-  overflow-x: scroll;
-  width: 100%;
-  border: 1px solid var(--table-wrapper-border-color);
-}
