+++ frontend/src/css/card.css
