+++ frontend/src/css/components/mb_data_table.css
